Use static_cast to silence sign conversion warning
This formula is perhaps non-standard, in that point size and pixels are
different measurements. Here they are being conflated as the same. We
may want to look into updating this in the future. For now, the explicit
static_cast marks that a type conversion is taking place here.

+++ NAS2D/Resources/Font.cpp
@@ -329,7 +329,7 @@ bool loadBitmap(const std::string& path, int glyphWidth, int glyphHeight, int gl
 
 	// Add generated texture id to texture ID map.
 	fontMap[path].texture_id = texture_id;
-	fontMap[path].pt_size = glyphHeight;
+	fontMap[path].pt_size = static_cast<unsigned int>(glyphHeight);
 	fontMap[path].height = glyphHeight;
 	fontMap[path].ref_count++;
 	fontMap[path].glyph_size = {glyphWidth, glyphHeight};
